Migration files issue fixed

+++ database/migrations/2022_05_30_145854_create_categories_table.php
@@ -16,8 +16,9 @@ return new class extends Migration
         Schema::create('categories', function (Blueprint $table) {
             $table->id();
             $table->string('category_name')->unique();
-            $table->unsignedInteger('posts_count')->default(0);
+            $table->unsignedBigInteger('posts_count')->default(0);
             $table->timestamps();
+
         });
     }
